+++ CHANGELOG.md
@@ -1,18 +1,34 @@
 # Changelog
+
+## [Unreleased]
+
+### Fixed
+
+- Diff URL did contain a wrong .git part
+
 ## [1.0.3] - 2024-04-05
+
 ### Added
+
 - Pretty print changelog
 
 ## 1.0.2 - 2024-04-05
+
 ### Added
+
 - Mocks
 
 ## 1.0.1 - 2024-04-05
+
 ### Added
+
 - Add Release to public API
 
 ## 1.0.0 - 2024-04-05
+
 ### Added
+
 - Initial version
 
+[Unreleased]: https://github.com/inlavigo/gg_changelog.git/compare/1.0.3...HEAD
 [1.0.3]: https://github.com/inlavigo/gg_changelog.git/compare/1.0.2...1.0.3